requirements.txt
# Machine Learning
torch>=2.9.0
torchvision>=0.20.0
segmentation-models-pytorch>=0.3.3
datasets>=2.16.1

# Geospatial
earthengine-api>=0.1.389
geopandas>=0.14.1
rasterio>=1.3.9
shapely>=2.0.2
pyproj>=3.6.1
geojson>=3.2.0

# Image Processing
opencv-python>=4.8.1
scikit-image>=0.22.0
Pillow>=10.1.0
albumentations>=1.3.1

# Utilities
numpy>=1.24.3
pandas>=2.1.3
pyyaml>=6.0.1
tqdm>=4.66.1
matplotlib>=3.8.2
